+++ CHANGELOG.md
@@ -11,3 +11,9 @@
 
 ### 1.0.3
 * Adicionado o pacotes api_error, api_security, status, params
+
+### 1.0.4
+* Alterado estrutura de pacotes
+
+### 1.0.5
+* Implementado adaptadores mongo e mysql